feat: use double instead of float for better precision
there is no real overhead in x86 architecture, but we will have to make some benchmark on other platforms

+++ include/components/mass.hpp
@@ -1,9 +1,8 @@
 #pragma once
 
 namespace components {
-    struct Mass
-    {
-            float mantissa;
+    struct Mass {
+            double mantissa;
             int exponent;
     };
 } // namespace components
